added check access

+++ src/main/java/ru/skypro/homework/service/impl/CommentServiceImpl.java
@@ -1,20 +1,25 @@
 package ru.skypro.homework.service.impl;
 
 import org.springframework.security.core.Authentication;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.skypro.homework.dto.CommentDTO;
 import ru.skypro.homework.dto.CreateCommentDTO;
 import ru.skypro.homework.dto.ResponseWrapperComment;
+import ru.skypro.homework.entity.Ad;
 import ru.skypro.homework.entity.Comment;
+import ru.skypro.homework.exception.AdNotFoundException;
 import ru.skypro.homework.mapper.CommentMapper;
 import ru.skypro.homework.repository.AdRepository;
 import ru.skypro.homework.repository.CommentRepository;
 import ru.skypro.homework.repository.UserRepository;
+import ru.skypro.homework.security.MyUserDetails;
 import ru.skypro.homework.service.CommentService;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @Transactional
@@ -24,12 +29,14 @@ public class CommentServiceImpl implements CommentService {
     private final AdRepository adRepository;
     private final UserRepository userRepository;
     private final CommentMapper commentMapper;
+    private final MyUserDetails userDetails;
 
-    public CommentServiceImpl(CommentRepository commentRepository, AdRepository adRepository, UserRepository userRepository, CommentMapper commentMapper) {
+    public CommentServiceImpl(CommentRepository commentRepository, AdRepository adRepository, UserRepository userRepository, CommentMapper commentMapper, MyUserDetails userDetails) {
         this.commentRepository = commentRepository;
         this.adRepository = adRepository;
         this.userRepository = userRepository;
         this.commentMapper = commentMapper;
+        this.userDetails = userDetails;
     }
 
     @Override
@@ -43,7 +50,8 @@ public class CommentServiceImpl implements CommentService {
     @Override
     public CommentDTO addComment(int id, CreateCommentDTO comment, Authentication authentication) {
         Comment newComment = commentMapper.toComment(comment);
-        newComment.setAd(adRepository.findById((long) id).get());
+        newComment.setAd(adRepository.findById((long) id)
+                .orElseThrow(AdNotFoundException::new));
         newComment.setCreatedAt(LocalDateTime.now());
         newComment.setAuthor(userRepository.getUserByEmail(authentication.getName()));
         commentRepository.save(newComment);
@@ -53,14 +61,33 @@ public class CommentServiceImpl implements CommentService {
 
 
     public void deleteComment(int adId, int commentId) {
-        commentRepository.deleteByIdAndAdId(commentId, adId);
+        if(checkAccess(commentRepository.findByIdAndAd_Id(commentId, adId))) {
+            commentRepository.deleteByIdAndAdId(commentId, adId);
+        }
+    }
+
+    @Override
+    public void deleteAllByAdId(long adId) {
+        commentRepository.deleteAllByAdId(adId);
     }
 
     @Override
     public CommentDTO updateComment(int adId, int commentId, CommentDTO commentDTO) {
         Comment updatedComment = commentRepository.findByIdAndAd_Id(commentId, adId);
+        if(checkAccess(updatedComment)){
         updatedComment.setText(commentDTO.getText());
         commentRepository.save(updatedComment);
+        }
         return commentMapper.toCommentDTO(updatedComment);
+    }
+
+    private boolean checkAccess(Comment comment){
+        if(userDetails.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_ADMIN"))) return true;
+        else{
+            Long commentAuthorId = comment.getAuthor().getId();
+            Long userId = Long.valueOf(userDetails.getIdUserDto());
+
+            return Objects.equals(userId, commentAuthorId);
+        }
     }
 }
